fix: make memory safety generic for all hardware and all architectures
Previous fix was Qwen3.5-specific. Now fully generic:

- _check_memory_safety() inspects the actual model at runtime to detect
  recurrent/hybrid layers (DeltaNet, SSM, Mamba) that use float32
  internally. Applies a memory_multiplier automatically — no hardcoded
  model_type checks.
- Auto-enables gradient checkpointing for ANY model when estimated
  memory > 80% of budget (was only Qwen3.5 before).
- Auto-reduces batch_size for ANY model when > 90% of budget, with
  compensating grad_accumulation_steps.
- Works on all Apple Silicon: M1 8GB, M2 Pro 16GB, M3 Pro 36GB,
  M4 Pro 48GB, etc. Budget is always 75% of system RAM.
- Fixed gated_delta_recurrence() (sequential path) which had the same
  full-tensor float32 upcast bug as the chunked path.

+++ mlx_forge/models/architectures/qwen3_5.py
@@ -185,17 +185,15 @@ def gated_delta_recurrence(
     B, T, Hk, Dk = q.shape
     Hv, Dv = v.shape[-2:]
 
-    # Upcast everything to float32 for numerical stability
-    q = q.astype(mx.float32)
-    k = k.astype(mx.float32)
-    v = v.astype(mx.float32)
+    # NOTE: Do NOT upcast full Q/K/V to float32 here — upcast per-step instead.
+    # Full upcast would allocate ~384 MB/layer × 21 layers = 8 GB on Qwen3.5.
 
     if state is None:
         state = mx.zeros((B, Hv, Dv, Dk), dtype=mx.float32)
     else:
         state = state.astype(mx.float32)
 
-    # Compute gates (already float32 internally)
+    # Compute gates in float32 (small: B × T × Hv)
     beta = mx.sigmoid(b.astype(mx.float32))  # (B, T, Hv)
     g = _compute_decay(A_log, a, dt_bias)  # (B, T, Hv) — float32
 
@@ -207,8 +205,11 @@ def gated_delta_recurrence(
 
     outputs = []
     for t in range(T):
+        # Per-step upcast to float32 (tiny: B × H × D per tensor)
         y, state = _gated_delta_step(
-            q[:, t], k[:, t], v[:, t],
+            q[:, t].astype(mx.float32),
+            k[:, t].astype(mx.float32),
+            v[:, t].astype(mx.float32),
             g[:, t], beta[:, t], state,
         )
         outputs.append(y)
